#530: Move recipe endpoints to correct location
