--- Api/Xipona.Api.Endpoints.Tests/v1/Controllers/RecipeControllerTests/GetRecipeByIdTests.cs
@@ -1,86 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.HttpResults;
-using ProjectHermes.Xipona.Api.ApplicationServices.Recipes.Queries.RecipeById;
-using ProjectHermes.Xipona.Api.Contracts.Common;
-using ProjectHermes.Xipona.Api.Contracts.Recipes.Queries.Get;
-using ProjectHermes.Xipona.Api.Domain.Common.Reasons;
-using ProjectHermes.Xipona.Api.Domain.Recipes.Models;
-using ProjectHermes.Xipona.Api.Domain.Recipes.Services.Queries;
-using ProjectHermes.Xipona.Api.Endpoint.v1.Controllers;
-using ProjectHermes.Xipona.Api.Endpoints.Tests.Common;
-using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
-using ProjectHermes.Xipona.Api.TestTools.Exceptions;
-
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.RecipeControllerTests;
-
-public class GetRecipeByIdTests : EndpointQueryNoConverterTestsBase<RecipeByIdQuery,
-    RecipeReadModel, RecipeContract, GetRecipeByIdTests.GetRecipeByIdFixture>
-{
-    public GetRecipeByIdTests() : base(new GetRecipeByIdFixture())
-    {
-    }
-
-    [Theory]
-    [InlineData(ErrorReasonCode.RecipeNotFound)]
-    public async Task EndpointCall_WithDomainException_ShouldReturnNotFound(ErrorReasonCode errorCode)
-    {
-        // Arrange
-        Fixture.SetupParameters();
-        Fixture.SetupQuery();
-        Fixture.SetupDomainException(errorCode);
-        Fixture.SetupDomainExceptionInQueryDispatcher();
-        Fixture.SetupExpectedErrorContract();
-        Fixture.SetupErrorConversion();
-
-        // Act
-        var result = await Fixture.ExecuteTestMethod();
-
-        // Assert
-        result.Should().BeOfType<NotFound<ErrorContract>>();
-        var notFound = result as NotFound<ErrorContract>;
-        notFound!.Value.Should().BeEquivalentTo(Fixture.ExpectedErrorContract);
-    }
-
-    public sealed class GetRecipeByIdFixture : EndpointQueryNoConverterFixtureBase
-    {
-        private Guid? _recipeId;
-
-        public GetRecipeByIdFixture()
-        {
-            PossibleResultsList.Add(new UnprocessableEntityStatusResult(ErrorReasonCode.RecipeNotFound));
-            PossibleResultsList.Add(new OkStatusResult());
-            PossibleResultsList.Add(new NotFoundStatusResult());
-        }
-
-        public override string RoutePattern => "/v1/recipes/{id:guid}";
-
-        public override async Task<IResult> ExecuteTestMethod()
-        {
-            TestPropertyNotSetException.ThrowIfNull(_recipeId);
-
-            return await RecipeEndpoints.GetRecipeById(
-                _recipeId.Value,
-                QueryDispatcherMock.Object,
-                ContractConverterMock.Object,
-                ErrorConverterMock.Object,
-                default);
-        }
-
-        public override void SetupParameters()
-        {
-            _recipeId = Guid.NewGuid();
-        }
-
-        public override void RegisterEndpoints(WebApplication app)
-        {
-            app.RegisterRecipeEndpoints();
-        }
-
-        public override void SetupQuery()
-        {
-            TestPropertyNotSetException.ThrowIfNull(_recipeId);
-            Query = new RecipeByIdQuery(new RecipeId(_recipeId.Value));
-        }
-    }
-}
