+++ Api/Xipona.Api.Endpoints.Tests/v1/Endpoints/RecipeTests/GetRecipeByIdTests.cs
@@ -1,30 +1,28 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
-using ProjectHermes.Xipona.Api.ApplicationServices.Recipes.Queries.ItemAmountsForOneServing;
+using ProjectHermes.Xipona.Api.ApplicationServices.Recipes.Queries.RecipeById;
 using ProjectHermes.Xipona.Api.Contracts.Common;
-using ProjectHermes.Xipona.Api.Contracts.Recipes.Queries.GetItemAmountsForOneServing;
+using ProjectHermes.Xipona.Api.Contracts.Recipes.Queries.Get;
 using ProjectHermes.Xipona.Api.Domain.Common.Reasons;
 using ProjectHermes.Xipona.Api.Domain.Recipes.Models;
 using ProjectHermes.Xipona.Api.Domain.Recipes.Services.Queries;
-using ProjectHermes.Xipona.Api.Endpoint.v1.Controllers;
+using ProjectHermes.Xipona.Api.Endpoint.v1.Endpoints;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.RecipeControllerTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Endpoints.RecipeTests;
 
-public class GetItemAmountsForOneServingTests : EndpointQueryNoConverterTestsBase<ItemAmountsForOneServingQuery,
-    IEnumerable<ItemAmountForOneServing>, ItemAmountsForOneServingContract,
-    GetItemAmountsForOneServingTests.GetItemAmountsForOneServingTestsFixture>
+public class GetRecipeByIdTests : EndpointQueryNoConverterTestsBase<RecipeByIdQuery,
+    RecipeReadModel, RecipeContract, GetRecipeByIdTests.GetRecipeByIdFixture>
 {
-    public GetItemAmountsForOneServingTests() : base(new GetItemAmountsForOneServingTestsFixture())
+    public GetRecipeByIdTests() : base(new GetRecipeByIdFixture())
     {
     }
 
     [Theory]
     [InlineData(ErrorReasonCode.RecipeNotFound)]
-    [InlineData(ErrorReasonCode.ItemNotFound)]
     public async Task EndpointCall_WithDomainException_ShouldReturnNotFound(ErrorReasonCode errorCode)
     {
         // Arrange
@@ -40,23 +38,34 @@ public class GetItemAmountsForOneServingTests : EndpointQueryNoConverterTestsBas
 
         // Assert
         result.Should().BeOfType<NotFound<ErrorContract>>();
-        var unprocessableEntity = result as NotFound<ErrorContract>;
-        unprocessableEntity!.Value.Should().BeEquivalentTo(Fixture.ExpectedErrorContract);
+        var notFound = result as NotFound<ErrorContract>;
+        notFound!.Value.Should().BeEquivalentTo(Fixture.ExpectedErrorContract);
     }
 
-    public sealed class GetItemAmountsForOneServingTestsFixture : EndpointQueryNoConverterFixtureBase
+    public sealed class GetRecipeByIdFixture : EndpointQueryNoConverterFixtureBase
     {
         private Guid? _recipeId;
 
-        public GetItemAmountsForOneServingTestsFixture()
+        public GetRecipeByIdFixture()
         {
+            PossibleResultsList.Add(new UnprocessableEntityStatusResult(ErrorReasonCode.RecipeNotFound));
             PossibleResultsList.Add(new OkStatusResult());
             PossibleResultsList.Add(new NotFoundStatusResult());
-            PossibleResultsList.Add(new UnprocessableEntityStatusResult(ErrorReasonCode.RecipeNotFound,
-                ErrorReasonCode.ItemNotFound));
         }
 
-        public override string RoutePattern => "/v1/recipes/{id:guid}/item-amounts-for-one-serving";
+        public override string RoutePattern => "/v1/recipes/{id:guid}";
+
+        public override async Task<IResult> ExecuteTestMethod()
+        {
+            TestPropertyNotSetException.ThrowIfNull(_recipeId);
+
+            return await RecipeEndpoints.GetRecipeById(
+                _recipeId.Value,
+                QueryDispatcherMock.Object,
+                ContractConverterMock.Object,
+                ErrorConverterMock.Object,
+                default);
+        }
 
         public override void SetupParameters()
         {
@@ -68,22 +77,10 @@ public class GetItemAmountsForOneServingTests : EndpointQueryNoConverterTestsBas
             app.RegisterRecipeEndpoints();
         }
 
-        public override async Task<IResult> ExecuteTestMethod()
-        {
-            TestPropertyNotSetException.ThrowIfNull(_recipeId);
-            return await RecipeEndpoints.GetItemAmountsForOneServing(
-                _recipeId.Value,
-                QueryDispatcherMock.Object,
-                ContractConverterMock.Object,
-                ErrorConverterMock.Object,
-                default);
-        }
-
         public override void SetupQuery()
         {
             TestPropertyNotSetException.ThrowIfNull(_recipeId);
-
-            Query = new ItemAmountsForOneServingQuery(new RecipeId(_recipeId.Value));
+            Query = new RecipeByIdQuery(new RecipeId(_recipeId.Value));
         }
     }
 }
